Free C# library consisting of various util classes for common day-to-day operations.
This library can be donwloaded free of charge and be distributed,
further modified and used for commercial gain according to MIT licence
that is shipped with this project.

This source code requires .NET 4.0 Framework to be installed on the
target machine to be successfully compiled.

There is also attached compiled version of this library, that is shipped
with other required stuff plus xml schema for VS IntelliSense.

If you have any (positiive/negative) feedback, do not hesitate to
contact me at  contact@lukaszdabrowski.com